pyproject.toml
[project]
name = "spa-web-application"
version = "0.1.0"
description = ""
authors = [
    {name = "Alexandr Ivanov",email = "stuardryan14529@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.2.5,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-filter (>=25.1,<26.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
